shorturl/pyproject.toml
[project]
name = "shorturl"
version = "0.1.0"
description = "URL shortener"
readme = "README.md"
authors = [
    { name = "Aero", email = "aerocyberdev@proton.me" }
]
requires-python = ">=3.13"
dependencies = [
    "appwrite>=7.0.1",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "validators>=0.34.0",
]
